+++ frontend/.eslintrc.cjs
@@ -7,7 +7,7 @@ module.exports = {
     "plugin:react/recommended",
   ],
   parser: "@typescript-eslint/parser",
-  plugins: ["@typescript-eslint"],
+  plugins: ["@typescript-eslint", "react-refresh"],
   root: true,
   env: {
     browser: true,
@@ -15,6 +15,7 @@ module.exports = {
   },
   rules: {
     "react/react-in-jsx-scope": "off",
-    "react/prop-types": "off"
+    "react/prop-types": "off",
+    "react-refresh/only-export-components": "warn"
   },
 }
