+++ frontend/package.json
@@ -38,6 +38,7 @@
     "daisyui": "^2.51.6",
     "eslint": "^8.39.0",
     "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-refresh": "^0.3.5",
     "postcss": "^8.4.23",
     "prettier": "^2.8.8",
     "tailwindcss": "^3.3.2",
